hotfix cached broadcast (delete cache file from root)

+++ src/automated_scoring/distributed.py
@@ -41,7 +41,8 @@ class DistributedExperiment(Experiment):
         temp_file_broadcast: str = self.comm.bcast(temp_file, root=0)
         data = from_cache(temp_file_broadcast)
         self.barrier()
-        os.remove(temp_file_broadcast)
+        if self.is_root:
+            os.remove(temp_file_broadcast)
         return data
 
     def barrier(self) -> None:
